refactor(dashboard-sidebar): update icon paths in Dashboard and Sidebar components
