+++ src/features/Dashboard.jsx
@@ -5,15 +5,15 @@ import Sidebar from "../components/Sidebar";
 //Iconos
 
 // Lo mismo a corregir que en el login.jsx
-import logoSmall from "../assets/icons/Logo_TalentMatch_AI_Small.png";
-import iconPositionCreatedBlue from "../assets/icons/icon_position_created_blue.png";
-import iconVacantCreateBlue from "../assets/icons/icon_vacant_create_blue.png";
-import iconVacantActiveBlue from "../assets/icons/icon_vacant_active_blue.png";
+import logoSmall from "../assets/icons/Logo_TalentMatch_AI_Small.svg";
+import iconPositionCreatedBlue from "../assets/icons/icon_position_created_blue.svg";
+import iconVacantCreateBlue from "../assets/icons/icon_vacant_create_blue.svg";
+import iconVacantActiveBlue from "../assets/icons/icon_vacant_active_blue.svg";
 //Iconos de botones
-import iconEyeHistoryGray from "../assets/icons/icon_eye_history.png";
-import iconCreateCircleplus from "../assets/icons/icon_create_circleplus.png";
-import iconUploadCv from "../assets/icons/icon_upload_cv.png";
-import iconVacantCreateButtom from "../assets/icons/icon_vacant_create_buttom.png";
+import iconEyeHistoryGray from "../assets/icons/icon_eye_history.svg";
+import iconCreateCircleplus from "../assets/icons/icon_create_circleplus.svg";
+import iconUploadCv from "../assets/icons/icon_upload_cv.svg";
+import iconVacantCreateButtom from "../assets/icons/icon_vacant_create_buttom.svg";
 
 {
   /** 1. El Acertijo del Componente Faltante (Tu observación)
